requirements.txt
adal==0.4.6
applicationinsights==0.10.0
argcomplete==1.8.2
asn1crypto==0.22.0
azure-batch==3.0.0
azure-cli==2.0.12
azure-cli-acr==2.0.9
azure-cli-acs==2.0.11
azure-cli-appservice==0.1.11
azure-cli-batch==3.0.3
azure-cli-billing==0.1.3
azure-cli-cdn==0.0.6
azure-cli-cloud==2.0.7
azure-cli-cognitiveservices==0.1.6
azure-cli-command-modules-nspkg==2.0.1
azure-cli-component==2.0.6
azure-cli-configure==2.0.10
azure-cli-consumption==0.1.3
azure-cli-container==0.1.7
azure-cli-core==2.0.12
azure-cli-cosmosdb==0.1.11
azure-cli-dla==0.0.10
azure-cli-dls==0.0.11
azure-cli-feedback==2.0.6
azure-cli-find==0.2.6
azure-cli-interactive==0.3.7
azure-cli-iot==0.1.10
azure-cli-keyvault==2.0.8
azure-cli-lab==0.0.9
azure-cli-monitor==0.0.8
azure-cli-network==2.0.11
azure-cli-nspkg==3.0.1
azure-cli-profile==2.0.9
azure-cli-rdbms==0.0.5
azure-cli-redis==0.2.7
azure-cli-resource==2.0.11
azure-cli-role==2.0.9
azure-cli-sf==1.0.5
azure-cli-sql==2.0.8
azure-cli-storage==2.0.11
azure-cli-vm==2.0.11
azure-common==1.1.8
azure-datalake-store==0.0.12
azure-graphrbac==0.30.0rc6
azure-keyvault==0.3.4
azure-mgmt-authorization==0.30.0rc6
azure-mgmt-batch==4.0.0
azure-mgmt-billing==0.1.0
azure-mgmt-cdn==0.30.2
azure-mgmt-cognitiveservices==1.0.0
azure-mgmt-compute==2.0.0
azure-mgmt-consumption==0.1.0
azure-mgmt-containerinstance==0.1.0
azure-mgmt-containerregistry==0.3.1
azure-mgmt-datalake-analytics==0.1.6
azure-mgmt-datalake-nspkg==2.0.0
azure-mgmt-datalake-store==0.1.6
azure-mgmt-devtestlabs==2.0.0
azure-mgmt-dns==1.0.1
azure-mgmt-documentdb==0.1.3
azure-mgmt-iothub==0.2.2
azure-mgmt-keyvault==0.40.0
azure-mgmt-monitor==0.2.1
azure-mgmt-network==1.3.0
azure-mgmt-nspkg==2.0.0
azure-mgmt-rdbms==0.1.0
azure-mgmt-redis==1.0.0
azure-mgmt-resource==1.1.0
azure-mgmt-sql==0.6.0
azure-mgmt-storage==1.1.0
azure-mgmt-trafficmanager==0.30.0
azure-mgmt-web==0.32.0
azure-monitor==0.3.0
azure-multiapi-storage==0.1.0
azure-nspkg==2.0.0
azure-servicefabric==5.6.130
bcrypt==3.1.3
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
colorama==0.3.9
cryptography==2.0.3
futures==3.1.1
humanfriendly==4.4.1
idna==2.5
isodate==0.5.4
jmespath==0.9.3
keyring==10.4.0
mock==2.0.0
msrest==0.4.11
msrestazure==0.4.11
oauthlib==2.0.2
paramiko==2.2.1
pbr==3.1.1
prompt-toolkit==1.0.15
pyasn1==0.3.2
pycparser==2.18
pydocumentdb==2.2.0
Pygments==2.2.0
PyJWT==1.5.2
PyNaCl==1.1.2
pyOpenSSL==17.2.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requests==2.18.3
requests-oauthlib==0.8.0
scp==0.10.2
SecretStorage==2.3.1
six==1.10.0
sshtunnel==0.1.2
tabulate==0.7.7
urllib3==1.22
vsts-cd-manager==0.118.0
wcwidth==0.1.7
Whoosh==2.7.4
xmltodict==0.11.0
